# laravel/source/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
        "/api/category": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get All Category",
                "operationId": "174eaa7695390e7c34bc0ac68fa70ea0",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Add Category",
                "operationId": "c9ec0a8cfdaf43141fee305b040272a3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "visible": {
                                        "type": "integer"
                                    },
                                    "text": {
                                        "type": "string"
                                    },
                                    "img": {
                                        "type": "string"
                                    },
                                    "parentsId": {
                                        "type": "integer"
                                    },
                                    "deleted": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/searchCategory/{name}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Tìm kiếm danh mục theo tên",
                "operationId": "0851f423ac506cc8fcbd99f4b13b0b52",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Tìm kiếm danh mục theo tên",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/category/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get Category",
                "operationId": "f0c3183fe0c45019dcd75f9529660bf6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The event ID specific to this event",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Category"
                ],
                "summary": "Update Category",
                "operationId": "a592d77165948f82c9e0eb551448bf54",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "visible": {
                                        "type": "integer"
                                    },
                                    "text": {
                                        "type": "string"
                                    },
                                    "img": {
                                        "type": "string"
                                    },
                                    "parentsId": {
                                        "type": "integer"
                                    },
                                    "deleted": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "Delete Category",
                "operationId": "d39845ee81a52b82283e3bdebaecec07",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The event ID specific to this event",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
        "/api/color": {
            "get": {
                "tags": [
                    "Color"
                ],
                "summary": "Get All Color",
                "operationId": "c5c78438dd88d26150442c9b275e0356",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Color"
                ],
                "summary": "Add Category",
                "operationId": "ba095fa7acd4e695631e7c0752cada95",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/color/{id}": {
            "get": {
                "tags": [
                    "Color"
                ],
                "summary": "Get Color",
                "operationId": "f7a5cbf38867eeaaae7833223fc50157",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Color"
                ],
                "summary": "Update Category",
                "operationId": "ea0c7d63dc1a77bd488ce7611920866c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Color"
                ],
                "summary": "Delete Color",
                "operationId": "2b35f197994112aa896e599d1ee4dda4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    }
}